remove unused code, refactor coloring
